Move dex to model/dex.go. Small fixes.
