+++ core/controlplane/config/config.go
@@ -88,16 +88,16 @@ func NewDefaultCluster() *Cluster {
 			},
 		},
 		Taints: []Taint{},
-		Dex: Dex{
+		Dex: model.Dex{
 			Enabled:         false,
 			Url:             "https://dex.example.com",
 			ClientId:        "example-app",
 			Username:        "email",
 			Groups:          "groups",
-			CaFile:          "",
-			Connectors:      []Connector{},
-			StaticClients:   []StaticClient{},
-			StaticPasswords: []StaticPassword{},
+			CaFile:          "/etc/kubernetes/ssl/openid-ca.pem",
+			Connectors:      []model.Connector{},
+			StaticClients:   []model.StaticClient{},
+			StaticPasswords: []model.StaticPassword{},
 		},
 	}
 
@@ -679,7 +679,7 @@ type Experimental struct {
 	NodeDrainer                 NodeDrainer              `yaml:"nodeDrainer"`
 	NodeLabels                  NodeLabels               `yaml:"nodeLabels"`
 	Plugins                     Plugins                  `yaml:"plugins"`
-	Dex                         Dex                      `yaml:"dex"`
+	Dex                         model.Dex                `yaml:"dex"`
 	DisableSecurityGroupIngress bool                     `yaml:"disableSecurityGroupIngress"`
 	NodeMonitorGracePeriod      string                   `yaml:"nodeMonitorGracePeriod"`
 	Taints                      []Taint                  `yaml:"taints"`
@@ -744,39 +744,6 @@ type KubeResourcesAutosave struct {
 
 type NodeDrainer struct {
 	Enabled bool `yaml:"enabled"`
-}
-
-type Dex struct {
-	Enabled         bool             `yaml:"enabled"`
-	Url             string           `yaml:"url"`
-	ClientId        string           `yaml:"clientId"`
-	Username        string           `yaml:"username"`
-	Groups          string           `yaml:"groups,omitempty"`
-	CaFile          string           `yaml:"caFile,omitempty"`
-	Connectors      []Connector      `yaml:"connectors,omitempty"`
-	StaticClients   []StaticClient   `yaml:"staticClients"`
-	StaticPasswords []StaticPassword `yaml:"staticPasswords"`
-}
-
-type Connector struct {
-	Type   string            `yaml:"type"`
-	Id     string            `yaml:"id"`
-	Name   string            `yaml:"name"`
-	Config map[string]string `yaml:"config"`
-}
-
-type StaticClient struct {
-	Id           string `yaml:"id"`
-	RedirectURIs string `yaml:"redirectURIs"`
-	Name         string `yaml:"name"`
-	Secret       string `yaml:"secret"`
-}
-
-type StaticPassword struct {
-	Email    string `yaml:"email"`
-	Hash     string `yaml:"hash"`
-	Username string `yaml:"username"`
-	UserId   string `yaml:"userID"`
 }
 
 type NodeLabels map[string]string
