+++ core/controlplane/config/templates/cluster.yaml
@@ -1090,7 +1090,7 @@ addons:
 #       id: github
 #       name: GitHub
 #       config:
-#         ClientId: "your_client_id"
+#         clientId: "your_client_id"
 #         clientSecret: "your_client_secret"
 #         redirectURI: https://dex.example.com/callback
 #         org: your_organization
@@ -1102,11 +1102,11 @@ addons:
 #       secret: 'ZXhhbXBsZS1hcHAtc2VjcmV0'
 #
 #     staticPasswords:
-#     - email: 'admin@example.com'
+#     - email: "admin@example.com"
 #       # bcrypt hash of the string "password". You can use bcrypt-tool from CoreOS to generate the passwords.
-#       hash: '$2a$10$2b2cU8CPhOTaGrs1HRQuAueS7JTT5ZHsHSzYiFPm1leZck7Mc8T4W'
-#       username: 'admin'
-#       userID: '08a8684b-db88-4b73-90a9-3cd1661f5466'
+#       hash: "$2a$10$2b2cU8CPhOTaGrs1HRQuAueS7JTT5ZHsHSzYiFPm1leZck7Mc8T4W"
+#       username: "admin"
+#       userID: "08a8684b-db88-4b73-90a9-3cd1661f5466"
 #
 #   # Kubernetes node labels to be added to controller nodes
 #   nodeLabels:
